Extract out GetArtifactsException from BuildResultHelper
The BuildResultHelper abstraction is no longer used post-cleanup, and
will be deleted. The GetArtifactsException that currently is inside the
interface is used, so this change extracts it out into a separate class.

Bug:374906681

Test: existing tests

AOSP: 1b413f38e00ac9d4533b66012cec51bd81ff700b

# Conflicts:
#	MODULE.bazel.lock
