+++ base/src/com/google/idea/blaze/base/bazel/LocalInvoker.java
@@ -28,6 +28,7 @@ import com.google.idea.blaze.base.command.WorkspaceRootReplacement;
 import com.google.idea.blaze.base.command.buildresult.BuildEventProtocolUtils;
 import com.google.idea.blaze.base.command.buildresult.BuildResult;
 import com.google.idea.blaze.base.command.buildresult.BuildResultHelper;
+import com.google.idea.blaze.base.command.buildresult.GetArtifactsException;
 import com.google.idea.blaze.base.command.buildresult.BuildResultHelperBep;
 import com.google.idea.blaze.base.command.buildresult.bepparser.BuildEventStreamProvider;
 import com.google.idea.blaze.base.console.BlazeConsoleLineProcessorProvider;
@@ -228,13 +229,13 @@ public class LocalInvoker extends AbstractBuildInvoker {
   }
 
   private BuildEventStreamProvider getBepStream(File outputFile)
-      throws BuildResultHelper.GetArtifactsException {
+    throws GetArtifactsException {
     try {
       return BuildEventStreamProvider.fromInputStream(
           new BufferedInputStream(new FileInputStream(outputFile)));
     } catch (FileNotFoundException e) {
       logger.error(e);
-      throw new BuildResultHelper.GetArtifactsException(e.getMessage());
+      throw new GetArtifactsException(e.getMessage());
     }
   }
