+++ base/src/com/google/idea/blaze/base/command/buildresult/GetArtifactsException.java
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2025 The Bazel Authors. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.google.idea.blaze.base.command.buildresult;
+
+import com.google.idea.blaze.exception.BuildException;
+
+/**
+ * Indicates a failure to get artifact information
+ */
+public class GetArtifactsException extends BuildException {
+  public GetArtifactsException(Throwable cause) {
+    super(cause);
+  }
+
+  public GetArtifactsException(String message) {
+    super(message);
+  }
+
+  public GetArtifactsException(String message, Throwable cause) {
+    super(message, cause);
+  }
+}
